Small refactor in structure
