+++ src/main/java/application/Engine.java
@@ -1,10 +1,13 @@
 package application;
 
-import application.data.*;
+import application.data.money.amount.AmountOfMoney;
+import application.data.money.amount.AmountOfMoneyException;
+import application.data.money.bank.BankAccount;
+import application.data.money.wallet.Wallet;
+import application.services.EnterDataInterface;
 
 import java.math.BigDecimal;
 import java.util.ArrayList;
-import java.util.Scanner;
 
 public class Engine implements EnterDataInterface {
     private final ArrayList<AmountOfMoney> bankAccounts = new ArrayList<>();
