+++ src/main/java/application/data/money/amount/AmountOfMoneyException.java
@@ -1,4 +1,4 @@
-package application.data;
+package application.data.money.amount;
 
 public class AmountOfMoneyException extends Exception {
     AmountOfMoneyException(String message) {
