+++ src/main/java/application/data/money/amount/AmountOfMoneyInterface.java
@@ -1,0 +1,20 @@
+package application.data.money.amount;
+
+interface AmountOfMoneyInterface {
+    /**
+     * Do wczytawania poczatkowego stanu gotowki
+     */
+    void enterMoney();
+
+    /**
+     * Do zwracania stanu gotowki
+     */
+    String toString();
+
+    /**
+     * Do listowania kont i portfeli
+     *
+     * @return zwraca nazwe obiektu
+     */
+    String getName();
+}
