+++ src/main/java/application/data/money/bank/BankAccount.java
@@ -1,51 +1,23 @@
-package application.data;
+package application.data.money.bank;
+
+import application.data.money.amount.AmountOfMoney;
+import application.services.NumbersDataBase;
+import application.data.user.UserData;
+import application.services.EnterDataInterface;
 
 import java.math.BigDecimal;
 import java.util.Scanner;
 
-interface BankAccountInterface {
-
-    /**
-     * Do wczytywania pol klasy
-     */
-
-    void enterDataBankAccount();
-
-    /**
-     * Zwraca wszystkie data konta
-     */
-
-    String toString();
-}
-
-class BankAccountException extends Exception {
-
-    BankAccountException(String message) {
-        super(message);
-    }
-}
-
 public class BankAccount extends AmountOfMoney implements BankAccountInterface {
-    // bede tworzyc transakcje i dodawac tutaj liste transakcji. pomysle nad ich limitem
     private final static int accountNumberDigits = 26;
     private final static NumbersDataBase numbersDataBase = new NumbersDataBase(accountNumberDigits);
     private BigDecimal numberBankAccount;
     private UserData accountHolder;
 
-    /**
-     * ==================== KONSTRUKTORY =========================
-     */
-
     public BankAccount() {
         this.numberBankAccount = null;
         this.accountHolder = null;
     }
-
-    /** ======================= METODY ============================ */
-
-    /**
-     * ---------------- Numer konta bankowego --------------------
-     */
 
     private void setNumberBankAccount(BigDecimal numberBankAccount) throws BankAccountException {
         if (numberBankAccount.compareTo(new BigDecimal(0)) < 0) {
@@ -64,10 +36,6 @@ public class BankAccount extends AmountOfMoney implements BankAccountInterface {
         return numberBankAccount;
     }
 
-    /**
-     * --------------------- Wlasciciel --------------------------
-     */
-
     public boolean hasAccountHolder() {
         // it gives information about existence of account holder in current bank account, concurrently avoiding getter
         return this.accountHolder != null;
@@ -82,12 +50,8 @@ public class BankAccount extends AmountOfMoney implements BankAccountInterface {
         } else System.out.println("Podpowiedz: Dane mozesz wpisac w kazdej chwili wybierajac opcje w menu startowym.");
     }
 
-    /**
-     * ---------- Wczytywanie danych konta bankowego -------------
-     */
-
     public void enterDataBankAccount() {
-        super.enterMoney();//wywoluje metode enterMoney klasy AmountOfMoney
+        super.enterMoney();     //wywoluje metode enterMoney klasy AmountOfMoney
         boolean toContinue;
         do {
             try {
@@ -102,10 +66,6 @@ public class BankAccount extends AmountOfMoney implements BankAccountInterface {
 
         enterAccountHolder();
     }
-
-    /**
-     * ------------ Odczyt danych konta bankowego ---------------
-     */
 
     @Override
     public String toString() {
