+++ src/main/java/application/data/money/bank/BankAccountException.java
@@ -1,0 +1,7 @@
+package application.data.money.bank;
+
+public class BankAccountException extends Exception {
+    BankAccountException(String message) {
+        super(message);
+    }
+}
