+++ src/main/java/application/data/money/bank/BankAccountInterface.java
@@ -1,0 +1,13 @@
+package application.data.money.bank;
+
+public interface BankAccountInterface {
+    /**
+     * Do wczytywania pol klasy
+     */
+    void enterDataBankAccount();
+
+    /**
+     * Zwraca wszystkie data konta
+     */
+    String toString();
+}
