+++ src/main/java/application/data/money/wallet/WalletInterface.java
@@ -1,0 +1,14 @@
+package application.data.money.wallet;
+
+interface WalletInterface {
+    /**
+     * prawdo podobnie bd wyorzystane do
+     * wczytywanie danych do portfela
+     */
+    void enterDataWallet();
+
+    /**
+     * odczyt danych portfela
+     */
+    String toString();
+}
