+++ src/test/java/application/data/BankAccountTest.java
@@ -1,5 +1,6 @@
 package application.data;
 
+import application.data.money.bank.BankAccount;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
