+++ src/test/java/application/data/EnterDataTest.java
@@ -1,5 +1,6 @@
 package application.data;
 
+import application.services.EnterDataInterface;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
