+++ src/test/java/application/data/NumbersDataBaseTest.java
@@ -1,5 +1,6 @@
 package application.data;
 
+import application.services.NumbersDataBase;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
