+++ src/test/java/application/data/UserDataTest.java
@@ -1,5 +1,6 @@
 package application.data;
 
+import application.data.user.UserData;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
